requirements.txt
numpy==1.22.1
Pillow==9.0.0
torch==1.10.2
torchvision==0.11.3
typing_extensions==4.0.1
